// src/app/components/security-module/models/enum-permission.type.model.ts
/**
 * Modelo de Permisos
 * Define los permisos disponibles en la aplicación
 */
export enum EnumPermissionType {
  // Entidades
  VIEW_USERS = 'VIEW_USERS',
  CREATE_USER = 'CREATE_USER',
  EDIT_USER = 'EDIT_USER',
  DELETE_USER = 'DELETE_USER',

  // Entidades
  VIEW_ENTITIES = 'VIEW_ENTITIES',
  CREATE_ENTITY = 'CREATE_ENTITY',
  EDIT_ENTITY = 'EDIT_ENTITY',
  DELETE_ENTITY = 'DELETE_ENTITY',
  OPEN_ENTITY = 'OPEN_ENTITY',

  // Personas
  VIEW_PERSONS = 'VIEW_PERSONS',
  CREATE_PERSON = 'CREATE_PERSON',
  EDIT_PERSON = 'EDIT_PERSON',
  DELETE_PERSON = 'DELETE_PERSON',

  // Artículos
  VIEW_ARTICLES = 'VIEW_ARTICLES',
  CREATE_ARTICLE = 'CREATE_ARTICLE',
  EDIT_ARTICLE = 'EDIT_ARTICLE',
  DELETE_ARTICLE = 'DELETE_ARTICLE',

  // Cotizaciones
  VIEW_QUOTES = 'VIEW_QUOTES',
  CREATE_QUOTE = 'CREATE_QUOTE',
  EDIT_QUOTE = 'EDIT_QUOTE',
  DELETE_QUOTE = 'DELETE_QUOTE',

  // Configuración
  VIEW_CONFIG = 'VIEW_CONFIG',
  EDIT_CONFIG = 'EDIT_CONFIG'
}